requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pytz==2017.2
SQLAlchemy==1.1.11
